+++ CHANGELOG.md
@@ -26,3 +26,8 @@ All notable changes to this project will be documented in this file.
 - Vertex.nodes supports conditions on edges
 - Vertex.in, Vertex.out Vertex.inE, Vertex.outE support depth as second parameter
 - IRuby-Support. Formatted output for Base-Objects
+
+## 0.5 - scheduled in dec. 2023
+- Arcade::Match simple match-statement generator
+- return to Arcade::QueryError messages instead of HTTPX::HttpError
+-